전역적으로 동작하는 PlayerFooterCell 구현
