+++ WeatherPlaylist/WeatherPlaylist/Views/StartView/RootView.swift
@@ -1,0 +1,21 @@
+//
+//  RootView.swift
+//  WeatherPlaylist
+//
+//  Created by 김태훈 on 12/12/23.
+//
+
+import SwiftUI
+
+struct RootView: View {
+    var body: some View {
+        VStack {
+            MainPageView()
+            PlayFooterCell()
+        }
+    }
+}
+
+#Preview {
+    RootView()
+}
